+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+ * Fix potential issues with `Parser` repeating the `@ByPtrPtr` or `@ByPtrRef` annotations on parameters
+ * To support Scala singleton objects better, consider as `static` methods from objects that are not `Pointer`
+ * Allow `Loader.extractResource()` and `cacheResource()` to extract or cache all files from a directory in a JAR file
+ * Create version-less symbolic links to libraries in cache on those platforms where it is useful to link easily
+ * Use `java.io.tmpdir` as fallback in `Loader.getCacheDir()`, and throw a clear exception on failure
+
 ### December 7, 2016 version 1.3
  * Print memory sizes in a human-readable format with `Pointer.formatBytes()`
  * Map standard `malloc()`, `calloc()`, `realloc()`, and `free()` functions ([issue #136](https://github.com/bytedeco/javacpp/issues/136))